Fix Surface_sweep

+++ Arrangement_on_surface_2/include/CGAL/Arr_geometry_traits/Conic_arc_2.h
@@ -1750,7 +1750,7 @@ protected:
       // Having computed y, x is the single solution to the quadratic equation
       // above, and since its discriminant is 0, x is simply given by:
       x = -(nt_traits.convert(_t)*ys[i] + nt_traits.convert(_u)) /
-        nt_traits.convert(_two*_r);
+        nt_traits.convert(Integer(_two * _r));
 
       ps[i] = Point_2 (x, ys[i]);
     }
